refactor(telemetry): move realtime span manager to seperate file
